Check sorts commits by PR created_at/head commit
Previously this resource would sort the open PRs by their updated at
time. This is not a good thing to sort by because any change to the PR
in Github causes the updated_at to change. Because there was no
timestamp associated with the resource's metadata this meant that
duplicates would be sent back to Concourse in varying orders, which is
no good.

This change sorts PRs them based on the created_at of the PR except for
the case where the created_at of the PR is older than the head commit
date. Additionally, it includes the timestamp it used for the sort in
the metadata as "timestamp".

This means that once a timestamp is detected and provided in the call to
check it will return all PRs that were created since that timestamp in
addition to PRs that were created prior to that timestamp having a HEAD
commit that is newer than the timestamp.

It's notable that this change causes more API calls to be made on your
behalf, similar to how the [ci skip] feature, if used, would result in
1+N API calls being made. The calls here are 1 to get the open PRs and
then 1 per open PR to get its HEAD commit date.